--- Game/Game.java
@@ -1,13 +1,0 @@
-package Game;
-
-import java.util.LinkedList;
-
-public class Game {
-    /*Deck deck;
-    LinkedList<Player> players;
-    public Game(LinkedList<Player> players){
-        this.players = players;
-        deck = new Deck();
-    }*/
-
-}
