+++ Game/Shared/EnCode.java
